fix: crashes related to Keyboard in FreezeTagConfigMenu
In some kingdoms (e.g. in Sand, but not in Cascade) it always/sometimes crashed the game when closing the keyboard.

I assume that's because the Keyboard was on the SceneHeap which it shouldn't.
At least moving it to the game mode heap fixed it.

But then there was a memory leak that the game mode heap growed everytime the scene was reloaded.
That was because of some never freed memory in the Keyboard class.
I tried to implement destructors for `StageSceneStateServerConfig` and `FreezeTagConfigMenu` but that also crashed.
So I added new `clean()` methods instead.